// part-09/patientor/backend/src/utils/types.ts
export interface Diagnoses {
    code: string;
    name: string;
    latin?: string;
}

export enum Gender {
    Male = "male",
    Female = "female",
    Other = "other",
}

export interface Patient {
    id: string;
    name: string;
    dateOfBirth: string;
    ssn: string;
    gender: Gender;
    occupation: string;
}

export type PatientPrivate = Omit<Patient, "ssn">;

export type NewPatientData = Omit<Patient, "id">;